finalize resource structure
