+++ resource_trigger/utils.py
